refactor: product와 같은 패키지 구조로 변경
