+++ src/test/java/roomescape/domain/member/MemberRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.domain.member;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,8 +7,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import roomescape.domain.member.Member;
-import roomescape.domain.member.MemberRepository;
 import roomescape.support.fixture.MemberFixture;
 
 @DataJpaTest
